perf: load dm20 images from static table

+++ include/image_loader/dm/load_images_dm.h
@@ -6,7 +6,7 @@
 
 namespace bongocat::animation {
     struct animation_thread_context_t;
-    bongocat_error_t init_dm_anim(animation_thread_context_t& ctx, int anim_index, const assets::embedded_image_t& sprite_sheet_image, int sprite_sheet_cols, int sprite_sheet_rows);
+    bongocat_error_t init_dm_anim(animation_thread_context_t& ctx, size_t anim_index, const assets::embedded_image_t& sprite_sheet_image, int sprite_sheet_cols, int sprite_sheet_rows);
 
     BONGOCAT_NODISCARD created_result_t<dm_sprite_sheet_t> load_dm_sprite_sheet(const animation_thread_context_t& ctx, size_t index);
 }
