+++ include/image_loader/dmx/load_images_dmx.h
@@ -6,7 +6,7 @@
 
 namespace bongocat::animation {
     struct animation_thread_context_t;
-    bongocat_error_t init_dmx_anim(animation_thread_context_t& ctx, int anim_index, const assets::embedded_image_t& sprite_sheet_image, int sprite_sheet_cols, int sprite_sheet_rows);
+    bongocat_error_t init_dmx_anim(animation_thread_context_t& ctx, size_t anim_index, const assets::embedded_image_t& sprite_sheet_image, int sprite_sheet_cols, int sprite_sheet_rows);
 
-    [[nodiscard]] created_result_t<dm_sprite_sheet_t> load_dmx_sprite_sheet(const animation_thread_context_t& ctx, int index);
+    [[nodiscard]] created_result_t<dm_sprite_sheet_t> load_dmx_sprite_sheet(const animation_thread_context_t& ctx, size_t index);
 }
