+++ include/image_loader/misc/load_images_misc.h
@@ -7,9 +7,9 @@
 
 namespace bongocat::animation {
 struct animation_thread_context_t;
-bongocat_error_t init_misc_anim(animation_thread_context_t& ctx, int anim_index,
+bongocat_error_t init_misc_anim(animation_thread_context_t& ctx, size_t anim_index,
                                 const assets::embedded_image_t& sprite_sheet_image,
                                 const assets::custom_animation_settings_t& sprite_sheet_settings);
 
-BONGOCAT_NODISCARD created_result_t<custom_sprite_sheet_t> load_misc_sprite_sheet(const animation_thread_context_t& ctx, int index);
+BONGOCAT_NODISCARD created_result_t<custom_sprite_sheet_t> load_misc_sprite_sheet(const animation_thread_context_t& ctx, size_t index);
 }  // namespace bongocat::animation
