+++ include/image_loader/ms_agent/load_images_ms_agent.h
@@ -11,14 +11,14 @@ BONGOCAT_NODISCARD created_result_t<ms_agent_sprite_sheet_t>
 load_ms_agent_sprite_sheet(const config::config_t& config, const assets::embedded_image_t& sprite_sheet_image,
                            int sprite_sheet_cols, int sprite_sheet_rows);
 BONGOCAT_NODISCARD created_result_t<ms_agent_sprite_sheet_t>
-load_ms_agent_anim(const animation_thread_context_t& ctx, int anim_index, const assets::embedded_image_t& sprite_sheet_image,
+load_ms_agent_anim(const animation_thread_context_t& ctx, size_t anim_index, const assets::embedded_image_t& sprite_sheet_image,
                    int sprite_sheet_cols, int sprite_sheet_rows,
                    const assets::ms_agent_animation_indices_t& animation_data);
 created_result_t<ms_agent_sprite_sheet_t>
-init_ms_agent_anim(animation_thread_context_t& ctx, int anim_index, const assets::embedded_image_t& sprite_sheet_image,
+init_ms_agent_anim(animation_thread_context_t& ctx, size_t anim_index, const assets::embedded_image_t& sprite_sheet_image,
                    int sprite_sheet_cols, int sprite_sheet_rows,
                    const assets::ms_agent_animation_indices_t& animation_data);
 
 BONGOCAT_NODISCARD created_result_t<ms_agent_sprite_sheet_t> load_ms_agent_sprite_sheet(const animation_thread_context_t& ctx,
-                                                                                        int index);
+                                                                                        size_t index);
 }  // namespace bongocat::animation
