+++ src/image_loader/min_dm/load_images_min_dm.cpp
@@ -6,7 +6,7 @@
 #include "image_loader/base_dm/load_dm.h"
 
 namespace bongocat::animation {
-bongocat_error_t init_min_dm_anim(animation_thread_context_t& ctx, int anim_index,
+bongocat_error_t init_min_dm_anim(animation_thread_context_t& ctx, size_t anim_index,
                                   const assets::embedded_image_t& sprite_sheet_image, int sprite_sheet_cols,
                                   int sprite_sheet_rows) {
   using namespace assets;
