+++ src/image_loader/min_dm/min_dm_load_sprite_sheet.cpp
@@ -8,7 +8,7 @@
 #include "image_loader/min_dm/load_images_min_dm.h"
 
 namespace bongocat::animation {
-created_result_t<dm_sprite_sheet_t> load_min_dm_sprite_sheet(const animation_thread_context_t& ctx, int index) {
+created_result_t<dm_sprite_sheet_t> load_min_dm_sprite_sheet(const animation_thread_context_t& ctx, size_t index) {
   using namespace animation;
   using namespace assets;
   switch (index) {
